refactor(matching): extract MatchingVoteContent from MatchingVoteScreen
Extract Scaffold body into standalone MatchingVoteContent widget so it
can be embedded in Event Now Bar matching bottom sheet without a Scaffold.
MatchingVoteScreen now delegates to MatchingVoteContent.

Closes #658

Constraint: Must render identically to pre-refactor MatchingVoteScreen
Rejected: Pass widget builder callback | tighter coupling, harder to test
Confidence: high
Scope-risk: narrow
